requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
rich>=13.0.0
typer>=0.9.0
python-dotenv>=1.0.0

# AI/CLI tools
google-generativeai>=0.3.0
anthropic>=0.7.0
openai>=1.0.0
requests>=2.31.0

# File handling and utilities
pathlib2>=2.3.7
watchdog>=3.0.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: Git integration
gitpython>=3.1.0

# Optional: Scheduling
schedule>=1.2.0 